WMIOT/requirements.txt
blinker==1.4
click==6.7
Flask==1.0.2
Flask-DebugToolbar==0.10.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysql-connector-python==8.0.11
protobuf==3.6.0
six==1.11.0
Werkzeug==0.14.1
